api_backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-extensions==3.2.3
django-hosts==6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.1
psycopg2-binary==2.9.6
django-redis==5.3.0
channels>=4.0
channels-redis>=4.0
daphne>=4.0
beautifulsoup4==4.12.2
websockets==11.0.3
django-dbbackup==4.1.0
celery==5.3.6
django-celery-beat==2.7.0